move scanner graph_util to util/
