+++ depobs/website/views.py
@@ -15,7 +15,7 @@ from werkzeug.exceptions import BadRequest, NotFound
 
 from depobs.database import models
 from depobs.scanner import graph_traversal
-from depobs.scanner import graph_util
+from depobs.util import graph_util
 from depobs.website.celery_tasks import get_celery_tasks
 from depobs.worker import scoring
 from depobs.worker import validators
