+++ depobs/worker/scoring.py
@@ -16,7 +16,7 @@ from depobs.database.models import (
     PackageVersion,
 )
 from depobs.scanner.graph_traversal import node_dep_ids_iter
-import depobs.scanner.graph_util as graph_util
+from depobs.util import graph_util
 
 
 log = logging.getLogger(__name__)
